Release snipe-it-cert-auth
This commit implements a way to utilize a client certificate
authentication mechanism in Snipe-IT.

Snipe-IT does not create a new user on first login by default. This is
where this crate comes into play: It parses a given client
certificate and checks if a matching user exists in the Snipe-IT
database. If such a user does not exist, it creates a new user via the
Snipe-IT REST API.